requirements.txt
Pillow==7.1.1 # Manange pictures
pyssim==0.4 # To compare if new created picture is like the original
piexif==1.1.3 # Save new created picture with the same EXIF information
